cart limited count product

+++ app/Http/Controllers/cartcontroller.php
@@ -16,7 +16,7 @@ class cartcontroller extends Controller
         $cart_items = cart::where('user_id', $user_id)->get();
 
 
-        return view('cart', compact('cart_items') );
+        return view('cart', compact('cart_items'));
     }
 
     public function add($product_id)
@@ -41,27 +41,30 @@ class cartcontroller extends Controller
 
         return redirect()->route('cartIndex');
     }
+
     public function update(Request $request, $id)
     {
         $cart_item = cart::where('id', $id)
             ->where('user_id', Auth::user()->id)
             ->firstOrFail();
-
+        $product = Product::find($cart_item->product_id);
+        if ($request->input('count') > $product->count) {
+            return redirect()->back()->with('error', 'Sorry, there are only ' . $product->count . ' units of this product available.');
+        }
         if ($request->input('count') > 0) {
             $cart_item->count = $request->input('count');
             $cart_item->save();
         } else {
             $cart_item->delete();
         }
-
-        return redirect()->route('cartIndex');
+            return redirect()->route('cartIndex');
     }
+
 
 
     public function remove($id)
     {
-        cart::where('id',$id)->delete();
-
+        cart::where('id', $id)->delete();
         return redirect(url('/cart'));
     }
 }
